feat: widen score type from u32 to u64 (#23)
* feat: widen score type from u32 to u64 across all packages

Allows scores to exceed the u32 max (4,294,967,295), supporting games
that need larger score ranges.

---------

+++ packages/interfaces/src/structs/leaderboard.cairo
@@ -18,7 +18,7 @@ pub struct LeaderboardEntry {
     /// Unique identifier for the entry
     pub id: u64,
     /// Score value
-    pub score: u32,
+    pub score: u64,
 }
 
 /// Result of a leaderboard operation
